requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.118.0
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
openai==2.2.0
openai-whisper==20250625
pillow==11.3.0
PyAudio==0.2.14
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
regex==2025.9.18
requests==2.32.5
sniffio==1.3.1
SpeechRecognition==3.14.3
starlette==0.48.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
